pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-markdownx>=4.0.9",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.5",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "jupyter>=1.1.1",
    "notebook>=7.4.7",
    "requests>=2.32.5",
]
